[GITHUB-21] Removes JAVA_HOME from defaults
This setting is not needed as GoCD will use the system wide
version of Java.

Also updates to the newest Sansible Java role that switches to
OpenJDK.

During Travis tests the dreaded temp file issue was encountered
so world readable tmp files has been turned on for the tests only.